Move DataComponentProps to types file
