+++ code/utils/types.ts
@@ -32,3 +32,61 @@ export type AirtableResponse = {
 }
 export type JSONResponse = DataItem[]
 export type CSVResponse = DataItem[]
+
+// Component
+export interface DataComponentProps {
+    // Component layout
+    width: number
+    height: number
+
+    // Data props
+    dataSource: DataSource
+    apiUrl: string | null
+    apiResponseDataKey: string | null
+    airtableUrl: string | null
+    airtableImageSize: AirtableImageSize
+    dataSourceFileType: DataSourceFileType
+    jsonFileUrl: string | null
+    csvFileUrl: string | null
+    tsvFileUrl: string | null
+
+    // List items layout
+    direction: FlexDirection
+    verticalAlignment: FlexAlignment
+    verticalDistribution: FlexDistribution
+    columns: number
+    gap: number
+    wrap: FlexWrap
+    horizontalGap: number
+    verticalGap: number
+    listItem?: React.ReactNode
+    loadingState?: React.ReactNode
+    loadingDelay: number
+    emptyState?: React.ReactNode
+
+    // Scrolling
+    isScrollEnabled: boolean
+    isDragScrollEnabled: boolean
+    isWheelScrollEnabled: boolean
+
+    // Search functionality
+    isSearchEnabled: boolean
+    searchTerm: string
+
+    // Sorting
+    shouldSort: boolean
+    sortKey: SortKey
+    sortDirection: SortDirection
+
+    // Component modes
+    mode: ComponentMode
+
+    // HTTP
+    overrideHttpHeaders: boolean
+    httpAuthorizationHeader: string
+    httpHeaders: string[]
+
+    // Event handlers
+    onItemTap: (item) => void
+    onItemLongPress: (item) => void
+}
